move common icons and styles to a util file
