+++ js/components/home/header.jsx
@@ -12,7 +12,7 @@ export default function Header() {
         I said, "Then let that second be every second I breathe." <br />
         He said, "Only once". <br />
         I said, "Then let that once, last forever." <br />
-        The genie told me to shut the fuck up. <br />
+        The gennie told me to shut the fuck up. <br />
         <br />
         {/* This has nothing to do with the project. I put this here for decoration. */}
         神说，诸水之间要有空气。多放大悲咒，积阳德。每天放大悲咒，积积阳阳德。
