+++ js/components/home/pagination_boxes.jsx
@@ -56,5 +56,3 @@ function Box(props) {
     </div>
   );
 }
-
-// box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
