+++ js/utils/utils.js
@@ -9,3 +9,9 @@ export function getReqHeader() {
 export function capitalizeFirst(s) {
   return s.charAt(0).toUpperCase() + s.slice(1);
 };
+
+export function formatDate(dateStr) {
+  const options = { month: 'short', day: 'numeric', year: 'numeric' };
+  const date = new Date(dateStr);
+  return date.toLocaleDateString('en-US', options);
+};
